Refactor job module structure
- Moved phase and site components to a new model directory.
- Updated import paths accordingly in the main init file.
- Renamed job.py to job/job.py for better organization.
