+++ src/easydiffraction/__init__.py
@@ -5,11 +5,11 @@
 from easycrystallography.Components.Lattice import Lattice as Lattice
 from easycrystallography.Components.SpaceGroup import SpaceGroup as SpaceGroup
 
-from .components.phase import Phase as Phase
-from .components.phase import Phases as Phases
-from .components.site import Atoms as Atoms
-from .components.site import Site as Site
-from .job import DiffractionJob as Job
+from .job.model.phase import Phase as Phase
+from .job.model.phase import Phases as Phases
+from .job.model.site import Atoms as Atoms
+from .job.model.site import Site as Site
+from .job.job import DiffractionJob as Job
 from .utils import download_from_repository
 
 __all__ = ['Job', 'download_from_repository']
